# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: IQV Dictionary API
  version: 1.0.0
  description: |
    IQV Dictionary servislerinin API dokümantasyonu.

    Bu dokümantasyon, repodaki gerçek backend kodundan (routes/controllers/
    middleware/validation) çıkarılmıştır; kodda karşılığı olmayan hiçbir
    endpoint eklenmemiştir.

    ## API Summary

    {{API_SUMMARY}}

    ## Classification

    **INTERNAL APIs** — IQV Dictionary uygulamasının kendi iş
    süreçlerinde kullandığı dahili business API operasyonlarıdır
    (Dictionary terim CRUD'u + Kişi/Personel CRUD'u).

    **EXTERNAL APIs** — Dış sistemler veya harici istemciler tarafından
    tüketilebilen API operasyonlarıdır. Bu projede tek karşılığı, eski
    Node-RED entegrasyonu için tutulan `POST /list-dictionary` ve
    `POST /create-dictionary` uyumluluk uçlarıdır (`legacy.routes.ts`) —
    aynı Dictionary servisine delege eder, ikinci bir iş mantığı içermez.

    **SYSTEM APIs** — Kimlik doğrulama, sağlık kontrolü ve teknik
    altyapı servisleridir (`/api/v1/auth/*`, `/health`).

    Her operasyon **TOTAL APIs** grubunda VE kendi sınıflandırma grubunda
    görünür; bu bilinçli bir sunum tekrarıdır ancak API Summary
    sayımında operasyon mükerrer sayılmaz.

    Gruplar iki seviyelidir: kök grup (TOTAL / INTERNAL / EXTERNAL /
    SYSTEM) altında iş alanı (Health, Auth, Dictionary, People) yer alır.
    Sınıflandırma tek kaynaktan gelir: her operasyondaki
    `x-iqv-classification` ve `x-iqv-domain` alanları.

...

...

...

...

...

...
    ALMAZ):

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/me:
    get:
      tags: [Auth]
      x-iqv-classification: SYSTEM
      x-iqv-domain: Auth
      summary: Geçerli oturumu doğrular
      description: >-
        Elindeki token'ı backend'e doğrulatır ve güncel kullanıcı bilgisini
        döner. Frontend, korumalı bir sayfa render etmeden ÖNCE bu uca
        istek atar ("protected page flash" düzeltmesi). Zorunlu JWT.
      responses:
        '200':
          description: Token geçerli.
          content:
            application/json:
              schema:
                type: object
                required: [success, user]
                properties:
                  success: { type: boolean, example: true }
                  user:
                    $ref: '#/components/schemas/SafeUser'
        '401':
          description: Token eksik, geçersiz veya süresi dolmuş.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/dictionary:
    get:
      tags: [Dictionary]
      x-iqv-classification: INTERNAL
      x-iqv-domain: Dictionary
      summary: Sözlük terimlerini listeler
      description: >-
        Sayfalı/aranabilir terim listesi. `dictionary.read` **veya**
        `settings.read` izinlerinden biri yeterlidir (Ayarlar ekranı bu
        endpoint'i paylaşır). Zorunlu JWT.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: group
          in: query
          description: 'Gruba göre filtre (ör. "IQV OS AI", "Endüstriyel").'
          schema: { type: string }
        - name: subgroup
          in: query
          description: Alt gruba göre filtre.
          schema: { type: string }
      responses:
        '200':
          description: Başarılı.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DictionaryListResponse'
        '401':
          description: Yetkilendirme başlığı eksik/geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: dictionary.read/settings.read izni yok.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      tags: [Dictionary]
      x-iqv-classification: INTERNAL
      x-iqv-domain: Dictionary
      summary: Yeni terim oluşturur
      description: >-
        Yeni bir sözlük terimi oluşturur. `dictionary.create` **veya**
        `settings.update` izinlerinden biri yeterlidir (Ayarlar ekranı bu
        endpoint'i paylaşır). Aynı grup/alt grup içinde aynı İngilizce
        terim zaten varsa 409 döner. Zorunlu JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DictionaryCreateInput'
      responses:
        '201':
          description: Oluşturuldu.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionarySingleResponse' }
        '401':
          description: Yetkilendirme başlığı eksik/geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: dictionary.create/settings.update izni yok.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Aynı grup/alt grupta aynı İngilizce terim zaten var.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: Gönderilen alanlar eksik/geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/v1/dictionary/{id}:
    parameters:
      - $ref: '#/components/parameters/DictionaryId'
    get:
      tags: [Dictionary]
      x-iqv-classification: INTERNAL
      x-iqv-domain: Dictionary
      summary: Tek bir terimi getirir
      description: 'ID ile tek bir sözlük terimi döner. `dictionary.read` izni gerekir. Zorunlu JWT.'
      responses:
        '200':
          description: Başarılı.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionarySingleResponse' }
        '400':
          description: Geçersiz kayıt kimliği.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Yetkilendirme başlığı eksik/geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: dictionary.read izni yok.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Kayıt bulunamadı.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    put:
      tags: [Dictionary]
      x-iqv-classification: INTERNAL
      x-iqv-domain: Dictionary
      summary: Terimi günceller (tam)
      description: 'Bir sözlük terimini günceller. `dictionary.update` izni gerekir. Zorunlu JWT.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DictionaryUpdateInput' }
      responses:
        '200':
          description: Güncellendi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionarySingleResponse' }
        '400':
          description: Geçersiz kayıt kimliği.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Yetkilendirme başlığı eksik/geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: dictionary.update izni yok.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Kayıt bulunamadı.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Aynı grup/alt grupta aynı İngilizce terim zaten var.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: Gönderilen alanlar geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    patch:
      tags: [Dictionary]
      x-iqv-classification: INTERNAL
      x-iqv-domain: Dictionary
      summary: Terimi günceller (kısmi)
      description: 'PUT ile aynı iş mantığı — kısmi güncelleme. `dictionary.update` izni gerekir. Zorunlu JWT.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DictionaryUpdateInput' }
      responses:
        '200':
          description: Güncellendi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionarySingleResponse' }
        '400':
          description: Geçersiz kayıt kimliği.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Yetkilendirme başlığı eksik/geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: dictionary.update izni yok.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Kayıt bulunamadı.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Aynı grup/alt grupta aynı İngilizce terim zaten var.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: Gönderilen alanlar geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      tags: [Dictionary]
      x-iqv-classification: INTERNAL
      x-iqv-domain: Dictionary
      summary: Terimi siler
      description: 'Bir sözlük terimini siler. `dictionary.delete` izni gerekir. Zorunlu JWT.'
      responses:
        '200':
          description: Silindi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeleteResponse' }
        '400':
          description: Geçersiz kayıt kimliği.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Yetkilendirme başlığı eksik/geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: dictionary.delete izni yok.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Kayıt bulunamadı.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /list-dictionary:
    post:
      tags: [Dictionary]
      x-iqv-classification: EXTERNAL
      x-iqv-domain: Dictionary
      summary: 'Sözlük terimlerini listeler (legacy Node-RED uyumluluk ucu)'
      description: >-
        **EXTERNAL API.** Eski Node-RED entegrasyonu için tutulan uyumluluk
        ucu — `GET /api/v1/dictionary` ile AYNI `DictionaryService`'i
        kullanır, ikinci bir iş mantığı içermez. Yeni entegrasyonlar
        `/api/v1/dictionary` kullanmalıdır. Zorunlu JWT.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              description: Liste sorgusu (query string yerine gövdede gönderilir).
              properties:
                page: { type: integer, example: 1 }
                limit: { type: integer, example: 20 }
                search: { type: string }
                group: { type: string }
                subgroup: { type: string }
      responses:
        '200':
          description: Başarılı.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryListResponse' }
        '401':
          description: Yetkilendirme başlığı eksik/geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /create-dictionary:
    post:
      tags: [Dictionary]
      x-iqv-classification: EXTERNAL
      x-iqv-domain: Dictionary
      summary: 'Yeni terim oluşturur (legacy Node-RED uyumluluk ucu)'
      description: >-
        **EXTERNAL API.** Eski Node-RED entegrasyonu için tutulan uyumluluk
        ucu — `POST /api/v1/dictionary` ile AYNI `DictionaryService`'i
        kullanır. Yeni entegrasyonlar `/api/v1/dictionary` kullanmalıdır.
        Zorunlu JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DictionaryCreateInput'
      responses:
        '201':
          description: Oluşturuldu.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionarySingleResponse' }
        '401':
          description: Yetkilendirme başlığı eksik/geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: Gönderilen alanlar eksik/geçersiz.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      description: 'Sayfa numarası (varsayılan: 1).'
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      description: 'Sayfa başına kayıt sayısı (üst sınır: 100).'
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Search:
      name: search
      in: query
      description: Serbest metin arama.
      schema: { type: string }
    DictionaryId:
      name: id
      in: path
      required: true
      description: Sözlük terimi Mongo ObjectId'si.
      schema: { type: string, example: 64f1c2b8e1a2b3c4d5e6f789 }
...
  schemas:
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 42 }
        totalPages: { type: integer, example: 3 }
    FieldError:
      type: object
      properties:
        field: { type: string, example: english_term }
        message: { type: string, example: İngilizce terim zorunludur. }
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: Kayıt bulunamadı. }
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
...
    DeleteResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { nullable: true, example: null }
...
    SafeUser:
      type: object
      description: Parola alanı hiçbir zaman API yanıtında yer almaz.
      properties:
        _id: { type: string }
        username: { type: string }
        full_name: { type: string }
        email: { type: string }
        role: { type: string, example: admin }
        permissions:
          type: array
          items: { type: string }
        company_id: { type: string }
        organization_id: { type: string }
        company_name: { type: string }
        status: { type: string, example: active }
...
    DictionaryTerm:
      type: object
      properties:
        _id: { type: string }
        english_term: { type: string }
        turkish_term: { type: string }
        description: { type: string }
        group: { type: string, example: IQV OS AI }
        subgroup: { type: string }
        created_by_id: { type: string }
        updated_by_id: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DictionaryCreateInput:
      type: object
      required: [english_term, turkish_term, description, group]
      properties:
        english_term: { type: string, maxLength: 200 }
        turkish_term: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
        group: { type: string, maxLength: 100 }
        subgroup: { type: string, maxLength: 100 }
    DictionaryUpdateInput:
      type: object
      description: Tüm alanlar opsiyoneldir; en az bir alan gönderilmelidir.
      properties:
        english_term: { type: string, maxLength: 200 }
        turkish_term: { type: string, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
        group: { type: string, maxLength: 100 }
        subgroup: { type: string, maxLength: 100 }
    DictionaryListResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items: { $ref: '#/components/schemas/DictionaryTerm' }
        pagination: { $ref: '#/components/schemas/Pagination' }
    DictionarySingleResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { $ref: '#/components/schemas/DictionaryTerm' }
...